moved verifySession to friend_functions
